[#175883590] add missing selectors & refactoring
