+++ ts/store/reducers/wallet/wallets.ts
@@ -10,17 +10,23 @@ import { getValueOrElse } from "../../../features/bonus/bpd/model/RemoteValue";
 import { abiSelector } from "../../../features/wallet/onboarding/store/abi";
 import {
   BancomatPaymentMethod,
-  enhancePaymentMethod,
+  BPayPaymentMethod,
+  CreditCardPaymentMethod,
   isBancomat,
+  isBPay,
+  isCreditCard,
   isRawCreditCard,
+  isSatispay,
   PaymentMethod,
   RawCreditCardPaymentMethod,
   RawPaymentMethod,
+  SatispayPaymentMethod,
   Wallet
 } from "../../../types/pagopa";
 
 import { PotFromActions } from "../../../types/utils";
 import { isDefined } from "../../../utils/guards";
+import { enhancePaymentMethod } from "../../../utils/paymentMethod";
 import { Action } from "../../actions/types";
 import { paymentUpdateWalletPsp } from "../../actions/wallet/payment";
 import {
@@ -95,7 +101,7 @@ export const getFavoriteWallet = (state: GlobalState) =>
 /**
  * @deprecated Using API v2 this selector is deprecated
  * If you are searching for credit card or bancomat use {@link creditCardWalletV1Selector}
- * - {@link pagoPaCreditCardWalletV1Selector} {@link bancomatSelector} instead
+ * - {@link pagoPaCreditCardWalletV1Selector} {@link bancomatListSelector} instead
  */
 export const walletsSelector = createSelector(
   getWallets,
@@ -143,12 +149,55 @@ export const paymentMethodsSelector = createSelector(
 );
 
 /**
+ * Return the
+ */
+export const rawCreditCardListSelector = createSelector(
+  [paymentMethodsSelector],
+  (
+    paymentMethods: pot.Pot<ReadonlyArray<RawPaymentMethod>, Error>
+  ): ReadonlyArray<RawCreditCardPaymentMethod> =>
+    pot.getOrElse(
+      pot.map(paymentMethods, w => w.filter(isRawCreditCard)),
+      []
+    )
+);
+
+/**
  * Return a bancomat list enhanced with the additional abi information
  */
-export const bancomatSelector = createSelector(
+export const bancomatListSelector = createSelector(
   [paymentMethodsSelector],
   (paymentMethodPot): pot.Pot<ReadonlyArray<BancomatPaymentMethod>, Error> =>
     pot.map(paymentMethodPot, paymentMethod => paymentMethod.filter(isBancomat))
+);
+
+/**
+ * Return a credit card list
+ */
+export const creditCardListSelector = createSelector(
+  [paymentMethodsSelector],
+  (paymentMethodPot): pot.Pot<ReadonlyArray<CreditCardPaymentMethod>, Error> =>
+    pot.map(paymentMethodPot, paymentMethod =>
+      paymentMethod.filter(isCreditCard)
+    )
+);
+
+/**
+ * Return a satispay list
+ */
+export const satispayListSelector = createSelector(
+  [paymentMethodsSelector],
+  (paymentMethodPot): pot.Pot<ReadonlyArray<SatispayPaymentMethod>, Error> =>
+    pot.map(paymentMethodPot, paymentMethod => paymentMethod.filter(isSatispay))
+);
+
+/**
+ * Return a BPay list
+ */
+export const bPayListSelector = createSelector(
+  [paymentMethodsSelector],
+  (paymentMethodPot): pot.Pot<ReadonlyArray<BPayPaymentMethod>, Error> =>
+    pot.map(paymentMethodPot, paymentMethod => paymentMethod.filter(isBPay))
 );
 
 /**
@@ -164,17 +213,6 @@ export const creditCardWalletV1Selector = createSelector(
         w =>
           w.paymentMethod && w.paymentMethod.walletType === WalletTypeEnum.Card
       )
-    )
-);
-
-export const creditCardSelector = createSelector(
-  [paymentMethodsSelector],
-  (
-    paymentMethods: pot.Pot<ReadonlyArray<RawPaymentMethod>, Error>
-  ): ReadonlyArray<RawCreditCardPaymentMethod> =>
-    pot.getOrElse(
-      pot.map(paymentMethods, w => w.filter(isRawCreditCard)),
-      []
     )
 );
