+++ ts/utils/paymentMethod.ts
@@ -1,23 +1,25 @@
 import { fromNullable } from "fp-ts/lib/Option";
-import {
-  isRawBancomat,
-  isRawBPay,
-  isRawCreditCard,
-  isRawSatispay,
-  RawBancomatPaymentMethod,
-  RawCreditCardPaymentMethod,
-  RawPaymentMethod
-} from "../types/pagopa";
+import { Abi } from "../../definitions/pagopa/walletv2/Abi";
+import bPayImage from "../../img/wallet/cards-icons/bPay.png";
+import pagoBancomatImage from "../../img/wallet/cards-icons/pagobancomat.png";
+import satispayImage from "../../img/wallet/cards-icons/satispay.png";
 import {
   cardIcons,
   getCardIconFromBrandLogo
 } from "../components/wallet/card/Logo";
-import pagoBancomatImage from "../../img/wallet/cards-icons/pagobancomat.png";
-import { IndexedById } from "../store/helpers/indexer";
-import { Abi } from "../../definitions/pagopa/walletv2/Abi";
 import I18n from "../i18n";
-import satispayImage from "../../img/wallet/cards-icons/satispay.png";
-import bPayImage from "../../img/wallet/cards-icons/bPay.png";
+import { IndexedById } from "../store/helpers/indexer";
+import {
+  BancomatPaymentMethod,
+  isRawBancomat,
+  isRawBPay,
+  isRawCreditCard,
+  isRawSatispay,
+  PaymentMethod,
+  RawBancomatPaymentMethod,
+  RawCreditCardPaymentMethod,
+  RawPaymentMethod
+} from "../types/pagopa";
 import { FOUR_UNICODE_CIRCLES } from "./wallet";
 
 export const getPaymentMethodHash = (
@@ -96,4 +98,35 @@ export const getTitleFromPaymentMethod = (
     );
   }
   return FOUR_UNICODE_CIRCLES;
+};
+
+export const enhanceBancomat = (
+  bancomat: RawBancomatPaymentMethod,
+  abiList: IndexedById<Abi>
+): BancomatPaymentMethod => ({
+  ...bancomat,
+  abiInfo: bancomat.info.issuerAbiCode
+    ? abiList[bancomat.info.issuerAbiCode]
+    : undefined,
+  caption: getTitleFromBancomat(bancomat, abiList),
+  icon: getImageFromPaymentMethod(bancomat)
+});
+
+export const enhancePaymentMethod = (
+  pm: RawPaymentMethod,
+  abiList: IndexedById<Abi>
+): PaymentMethod => {
+  switch (pm.kind) {
+    // bancomat need a special handling, we need to include the abi
+    case "Bancomat":
+      return enhanceBancomat(pm, abiList);
+    case "CreditCard":
+    case "BPay":
+    case "Satispay":
+      return {
+        ...pm,
+        caption: getTitleFromPaymentMethod(pm, abiList),
+        icon: getImageFromPaymentMethod(pm)
+      };
+  }
 };
